Seprate models created for followers and following
